refactor maven
